Branding: handle repo errors in cog
If we fail to fetch an event, the whole branding sync will
now be aborted. This will prevent situations where we fail
to fetch the current event due to a 5xx error and the cog
resorts to the fallback branding in the middle of an event.

Error handling is moved to the cog. The repo abstraction
will now propagate errors rather than silence them.